requirements.txt
ultralytics>=8.4.12
pymongo>=4.6.0
python-dotenv>=1.0.0
opencv-python>=4.13.0
mediapipe>=0.10.32
pygame>=2.5.0
django-mongodb-backend>=6.0.2s
elevenlabs>=2.34.0